refactor: adjusts standard angular structure
