+++ src/app/app.module.ts
@@ -2,13 +2,12 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { Router } from '@angular/router';
+
 //app
-import { AppComponent } from '../modules/app/app.component';
-import { PageNotFoundComponent } from '../modules/page-not-found/page-not-found.component';
-
+import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app.routing.module';
-import { HomeModule } from '../modules/home/home.module';
-
+import { HomeModule } from './modules/home/home.module';
+import { PageNotFoundComponent } from './modules/page-not-found/page-not-found.component';
 
 @NgModule({
   imports: [ BrowserModule, HomeModule, AppRoutingModule ],
