+++ src/app/app.routing.module.ts
@@ -1,9 +1,9 @@
 //core
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-//app
-import { PageNotFoundComponent } from '../modules/page-not-found/page-not-found.component';
 
+//app
+import { PageNotFoundComponent } from './modules/page-not-found/page-not-found.component';
 
 export const appRoutes: Routes = [
   { path: '', redirectTo: '/home', pathMatch: 'full' },
